third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/computed-style-font-expected.txt

This test exercises the 'font' shorthand property in CSS computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedFont('10px sans-serif') is '10px sans-serif'
PASS computedFont('10px sans-serif') is '10px sans-serif'
PASS computedFont('10px SANS-SERIF') is '10px sans-serif'
PASS computedFont('12px sans-serif') is '12px sans-serif'
PASS computedFont('12px  sans-serif') is '12px sans-serif'
PASS computedFont('10px sans-serif, sans-serif') is '10px sans-serif, sans-serif'
PASS computedFont('10px sans-serif, serif') is '10px sans-serif, serif'
PASS computedFont('12px ahem') is '12px ahem'
PASS computedFont('12px unlikely-font-name') is '12px unlikely-font-name'
PASS computedFont('100 10px sans-serif') is '100 10px sans-serif'
PASS computedFont('200 10px sans-serif') is '200 10px sans-serif'
PASS computedFont('300 10px sans-serif') is '300 10px sans-serif'
PASS computedFont('400 10px sans-serif') is '10px sans-serif'
PASS computedFont('normal 10px sans-serif') is '10px sans-serif'
PASS computedFont('500 10px sans-serif') is '500 10px sans-serif'
PASS computedFont('600 10px sans-serif') is '600 10px sans-serif'
PASS computedFont('700 10px sans-serif') is '700 10px sans-serif'
PASS computedFont('bold 10px sans-serif') is '700 10px sans-serif'
PASS computedFont('800 10px sans-serif') is '800 10px sans-serif'
PASS computedFont('900 10px sans-serif') is '900 10px sans-serif'
PASS computedFont('italic 10px sans-serif') is 'italic 10px sans-serif'
PASS computedFont('small-caps 10px sans-serif') is 'small-caps 10px sans-serif'
PASS computedFont('italic small-caps 10px sans-serif') is 'italic small-caps 10px sans-serif'
PASS computedFont('italic small-caps bold 10px sans-serif') is 'italic small-caps 700 10px sans-serif'
PASS computedFont('10px/100% sans-serif') is '10px / 10px sans-serif'
PASS computedFont('10px/100px sans-serif') is '10px / 100px sans-serif'
PASS computedFont('10px/normal sans-serif') is '10px sans-serif'
PASS computedFont('10px/normal sans-serif') is '10px sans-serif'
PASS successfullyParsed is true

TEST COMPLETE
